+++ frontend/src/app/modules/shared/assembler/parts.assembler.spec.ts
@@ -175,21 +175,4 @@ describe('PartsAssembler', () => {
         });
     });
   });
-
-  describe('assembleAssetsCountryMap', () => {
-    it('should handle invalid values', () => {
-      expect(PartsAssembler.assembleAssetsCountryMap(null)).toBe(null);
-      expect(PartsAssembler.assembleAssetsCountryMap(undefined)).toBe(null);
-    });
-
-    it('should return country map', () => {
-      expect(
-        PartsAssembler.assembleAssetsCountryMap({
-          DEU: 100,
-        }),
-      ).toEqual({
-        DEU: 100,
-      });
-    });
-  });
 });